+++ lang.py
@@ -1,0 +1,39 @@
+import streamlit as st
+import numpy as np
+import plotly.express as px
+
+def app():
+    st.title('APP2')
+    st.write('Welcome to app2')
+import streamlit as st
+import altair as alt
+import pandas as pd
+from add_data import db_execute_fetch
+
+def app():
+    st.title('Languages')
+    st.write('You can see the languages piechart ')
+    langPie()
+    st.markdown("<p style='padding:10px; background-color:#000000;color:#00ECB9;font-size:16px;border-radius:10px;'>Section Break</p>", unsafe_allow_html=True)
+
+def loadData():
+    query = "select * from TweetInformation"
+    df = db_execute_fetch(query, dbName="tweets", rdf=True)
+    return df
+
+def langPie():
+    df = loadData()
+    dfLangCount = pd.DataFrame({'Tweet_count': df.groupby(['language'])['clean_text'].count()}).reset_index()
+    dfLangCount["language"] = dfLangCount["language"].astype(str)
+    dfLangCount = dfLangCount.sort_values("Tweet_count", ascending=False)
+    dfLangCount.loc[dfLangCount['Tweet_count'] < 10, 'lang'] = 'Other languages'
+    st.title(" Tweets Language pie chart")
+    fig = px.pie(dfLangCount, values='Tweet_count', names='language', width=500, height=350)
+    fig.update_traces(textposition='inside', textinfo='percent+label')
+
+    colB1, colB2 = st.beta_columns([2.5, 1])
+
+    with colB1:
+        st.plotly_chart(fig)
+    with colB2:
+        st.write(dfLangCount)
